netcache: add fwd table configs, missing nc header port logic, moved to tofino folder.

+++ tofino/netcache/controller/src/packet.h
@@ -17,7 +17,7 @@ struct netcache_hdr_t {
 	uint8_t seq;
 	uint16_t key;
 	uint32_t val;
-	uint8_t cpu;
+	uint16_t cpu;
 } __attribute__((packed));
 
 };	// namespace netcache
